Move visual resources into assets package
