+++ zapzap/assets/__init__.py
@@ -1,0 +1,1 @@
+"""Visual assets for ZapZap."""
