+++ zapzap/assets/icons/SystemIcon.py
