+++ zapzap/assets/icons/TrayIcon.py
