+++ zapzap/assets/icons/__init__.py
@@ -1,0 +1,1 @@
+"""Icon assets and icon factories."""
