+++ zapzap/assets/themes/__init__.py
@@ -1,0 +1,1 @@
+"""Theme assets and style providers."""
