+++ zapzap/controllers/QtoasterDonation.py
@@ -4,7 +4,7 @@ from PyQt6 import QtCore, QtWidgets
 from PyQt6.QtGui import QDesktopServices
 from PyQt6.QtCore import QUrl
 from gettext import gettext as _
-from zapzap.resources.UserIcon import UserIcon
+from zapzap.assets.icons.UserIcon import UserIcon
 from zapzap.core.environment.EnvironmentDetector import EnvironmentDetector
 from zapzap.views.ui_qtoaster_donation import Ui_QtoasterDonation
 from zapzap.core.config.SettingsManager import SettingsManager
