+++ zapzap/features/notifications/FreedesktopNotificationBackend.py
@@ -9,7 +9,7 @@ from PyQt6.QtWidgets import QApplication
 from PyQt6.QtWebEngineCore import QWebEngineNotification
 
 from zapzap.webengine import WebView
-from zapzap.resources.TrayIcon import TrayIcon
+from zapzap.assets.icons.TrayIcon import TrayIcon
 from zapzap.core.config.SettingsManager import SettingsManager
 from zapzap import __appname__
