+++ zapzap/models/card_user_model.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 
 from zapzap.models.User import User
-from zapzap.resources.UserIcon import UserIcon
+from zapzap.assets.icons.UserIcon import UserIcon
 from zapzap.core.config.SettingsManager import SettingsManager
