+++ zapzap/views/browser/browser_page_button.py
@@ -6,7 +6,7 @@ from PyQt6.QtGui import QIcon
 from PyQt6.QtWidgets import QPushButton
 
 from zapzap.models import User
-from zapzap.resources.UserIcon import UserIcon
+from zapzap.assets.icons.UserIcon import UserIcon
 from zapzap.core.config.SettingsManager import SettingsManager
