+++ zapzap/views/settings/about_settings_view.py
@@ -3,7 +3,7 @@
 from gettext import gettext as _
 from PyQt6.QtCore import QSize
 
-from zapzap.resources.UserIcon import UserIcon
+from zapzap.assets.icons.UserIcon import UserIcon
 from zapzap.views.components import Button, Label
 from zapzap.views.settings_components import (
     SettingsActionRow,
